+++ packages/spacecat-shared-google-client/CHANGELOG.md
@@ -1,3 +1,17 @@
+# [@adobe/spacecat-shared-google-client-v1.2.11](https://github.com/adobe/spacecat-shared/compare/@adobe/spacecat-shared-google-client-v1.2.10...@adobe/spacecat-shared-google-client-v1.2.11) (2024-10-08)
+
+
+### Bug Fixes
+
+* gracefully handle api call in case site url is not defined ([#397](https://github.com/adobe/spacecat-shared/issues/397)) ([45924c8](https://github.com/adobe/spacecat-shared/commit/45924c8ef341b238647cb10ea927ab7c634007d2))
+
+# [@adobe/spacecat-shared-google-client-v1.2.10](https://github.com/adobe/spacecat-shared/compare/@adobe/spacecat-shared-google-client-v1.2.9...@adobe/spacecat-shared-google-client-v1.2.10) (2024-10-05)
+
+
+### Bug Fixes
+
+* **deps:** update external fixes ([#394](https://github.com/adobe/spacecat-shared/issues/394)) ([ad165f5](https://github.com/adobe/spacecat-shared/commit/ad165f5fcf9c0ab17d2e901e7a5c0ed3c39dcb51))
+
 # [@adobe/spacecat-shared-google-client-v1.2.9](https://github.com/adobe/spacecat-shared/compare/@adobe/spacecat-shared-google-client-v1.2.8...@adobe/spacecat-shared-google-client-v1.2.9) (2024-09-28)